Changed size field in OPENFILE struct to uint64_t. Changed size, location fields in FILESYSTEM struct to uint64_t. Removed duplicate kprintf() function. Changed references from kprintf() to kprintf_direct() in readkey(). Added HEAP_HEADER_CORRUPT error. Fixed error in heapfree() where it did not set an error code when a corrupt heap header was found. Removed duplicate errors from error.h. Renamed error END_OF_DIR to END_OF_DIRECTORY. Replaced END_OF_DIR error with END_OF_DIRECTORY in dir_statement(), fat_find_free_block(), fat_find(), fat_create_int(). Replaced INVALID_EXEC error with INVALID_EXECUTABLE in load_elf32(). Rewrote dir_statement() to use do...while loop. Rewrote code to find handles to make it more efficient in dup2(). Fixed error in exec() where it did not store the process's heap end in the PROCESS list. Replaced kernel data segment constant 0x10 with KERNEL_DATA_SELECTOR in switch_task().

+++ filemanager/vfs.h
@@ -41,8 +41,8 @@
 typedef struct {
 	uint8_t name[MAX_PATH];
 	uint8_t magicnumber[MAX_PATH];
-	size_t size;
-	size_t location;
+	uint64_t size;
+	uint64_t location;
 	size_t (*findfirst)(char *name,void *);	/* handlers */
 	size_t (*findnext)(char *name,void *);
 	size_t (*read)(size_t,void *,size_t);
@@ -71,7 +71,7 @@ typedef struct {
 	TIME create_time_date;
 	TIME last_written_time_date;
 	TIME last_accessed_time_date;
-	uint32_t filesize;
+	uint64_t filesize;
 	uint64_t startblock;
 	size_t drive;
 	uint64_t currentblock;
@@ -87,7 +87,7 @@ typedef struct {
 	uint64_t findlastblock; /* last block read by find() */
 	size_t findentry;
 	size_t owner_process;
-	uint8_t searchfilename[MAX_PATH]; 		
+	uint8_t searchfilename[MAX_PATH];
 	struct FILERECORD *next;
 } __attribute__((packed)) FILERECORD;
